==> 2 - AddSmiles.ipynb <==
### Imports

import pubchempy as pcp
from pubchempy import Compound, get_compounds
import pandas as pd
import numpy as np

### Cargamos el dataset

odor = pd.read_json('odor_relations.json', lines=False, orient='records')

odor.head()

odor.tail()

### Borramos unos cids que dan error

#Estos dan error, así que los borramos
odor=odor.drop(odor[(odor["cid"]=='163640541')].index)
odor=odor.drop(odor[(odor["cid"]=='Search')].index)

### Hacemos una lista de todos los smiles

smiles_all=[]
for i in odor["cid"]:
    c=Compound.from_cid(i)
    smiles_all.append(c.isomeric_smiles)

### Creamos el dataset y lo expotamos

odor.insert(3,"smiles",smiles_all,True)

odor.head()

odor.tail()

#comprobar que los cids y el smile etan en el orden correcto
Compound.from_cid(5367681).isomeric_smiles=='CC/C=C\CCOC(=O)CC(C)C'
Compound.from_cid(7358).isomeric_smiles=='CCC(CC)CO'

odor.to_json (r'C:\Users\nlupe\Desktop\QUART CURS\TFG\odor_smiles.json', lines=False, orient='records')
