==> 3 - AddDescriptors.ipynb <==
### Imports

import numpy as np
import pandas as pd
from rdkit import Chem
from rdkit.Chem import AllChem
from rdkit.Chem.Pharm2D import Generate
from rdkit.Chem import Descriptors
from rdkit.Chem import rdMolDescriptors
from rdkit.ML.Descriptors import MoleculeDescriptors
from rdkit.Chem.EnumerateStereoisomers import EnumerateStereoisomers
from rdkit.Chem.EnumerateStereoisomers import StereoEnumerationOptions

### Data

odor = pd.read_json('odor_smiles.json', lines=False, orient='records')

odor.head()

odor.tail()

### Calculamos los descriptores 2D

descriptors=["TPSA","MolLogP","MolWt", "ExactMolWt", "MolMR", "NumHAcceptors", "NumHDonors", "NumRotatableBonds",
"NumAromaticRings", "NumSaturatedRings", "NumAliphaticRings", "NumAromaticHeterocycles", "NumSaturatedHeterocycles",
"NumAliphaticHeterocycles", "NumAromaticCarbocycles", "NumSaturatedCarbocycles", "NumAliphaticCarbocycles", "NumHeteroatoms", 
"NumValenceElectrons", "RingCount", "FractionCSP3", "LabuteASA", "BalabanJ", "BertzCT", "Ipc", "HallKierAlpha", "Kappa1",
"Kappa2", "Kappa3", "Chi0", "Chi1", "Chi0n", "Chi1n", "Chi2n", "Chi3n", "Chi4n", "Chi0v", "Chi1v", "Chi2v", "Chi3v", "Chi4v"]

descriptor_molecule = pd.DataFrame(columns=descriptors)

descriptor_molecule.head()

smiles=odor['smiles']

calculator = MoleculeDescriptors.MolecularDescriptorCalculator(descriptors)
for i in range (len(smiles)):
    mol=Chem.MolFromSmiles(smiles[i])
    descriptor_molecule.loc[i]=calculator.CalcDescriptors(mol)

descriptor_molecule.head()

odor_descriptors_1=pd.concat([odor, descriptor_molecule], axis=1)

odor_descriptors_1

#Comprobar que se haya hecho bien
calculator = MoleculeDescriptors.MolecularDescriptorCalculator(descriptors)
mol=Chem.MolFromSmiles('CO')
print(calculator.CalcDescriptors(mol))

calculator = MoleculeDescriptors.MolecularDescriptorCalculator(descriptors)
mol=Chem.MolFromSmiles('C/C=C/C1=CC=C(C=C1)OC')
print(calculator.CalcDescriptors(mol))
#OK

#### Añadimos dos descriptores más

HeavyAtom=[]
Gasteiger=[]

for i in range (len(smiles)):
    mol=Chem.MolFromSmiles(smiles[i])
    AllChem.ComputeGasteigerCharges(mol)
    HeavyAtom.append(Chem.Lipinski.HeavyAtomCount(mol))
    Gasteiger.append(sum([mol.GetAtomWithIdx(i).GetDoubleProp('_GasteigerCharge') for i in range(mol.GetNumAtoms())]))

twocolumns = pd.DataFrame()
twocolumns = twocolumns.assign(HeavyAtom=None)
twocolumns = twocolumns.assign(Gasteiger=None)

twocolumns["HeavyAtom"]=HeavyAtom
twocolumns["Gasteiger"]=Gasteiger

odor_descriptors=pd.concat([odor_descriptors_1, twocolumns], axis=1)

odor_descriptors.head()

#Comprobar que se haya hecho bien

mol=Chem.MolFromSmiles('CCC(=O)CCC1C(=CCCC1(C)C)C.C#CCO')
AllChem.ComputeGasteigerCharges(mol)
print(Chem.Lipinski.HeavyAtomCount(mol), [mol.GetAtomWithIdx(i).GetDoubleProp('_GasteigerCharge') for i in range(mol.GetNumAtoms())])

mol=Chem.MolFromSmiles('CO')
AllChem.ComputeGasteigerCharges(mol)
print(Chem.Lipinski.HeavyAtomCount(mol), [mol.GetAtomWithIdx(i).GetDoubleProp('_GasteigerCharge') for i in range(mol.GetNumAtoms())])

#OK

### Calculamos los descriptores 3D

descriptors3d=["Eccentricity", "InertialShapeFactor", "RadiusOfGyration", "Asphericity", "SpherocityIndex", 
             "PMI1", "PMI2", "PMI3", "NPR1", "NPR2"]

r=len(odor['smiles'])
descriptor_molecule3d = pd.DataFrame(columns=descriptors3d, index=range(r))

descriptor_molecule3d

for i in range (len(smiles)):
    mol=Chem.AddHs(Chem.MolFromSmiles(smiles[i]))
    conformers = AllChem.EmbedMultipleConfs(mol, numConfs=1)
    E=0
    I=0
    R=0
    A=0
    S=0
    P1=0
    P2=0
    P3=0
    N1=0
    N2=0
    for j in range(100):
        try:
            E+=Chem.Descriptors3D.Eccentricity(mol)
            I+=Chem.Descriptors3D.InertialShapeFactor(mol)
            R+=Chem.Descriptors3D.RadiusOfGyration(mol)
            A+=Chem.Descriptors3D.Asphericity(mol)
            S+=Chem.Descriptors3D.SpherocityIndex(mol)
            P1+=Chem.Descriptors3D.PMI1(mol)
            P2+=Chem.Descriptors3D.PMI2(mol)
            P3+=Chem.Descriptors3D.PMI3(mol)
            N1+=Chem.Descriptors3D.NPR1(mol)
            N2+=Chem.Descriptors3D.NPR2(mol)
            
            descriptor_molecule3d.iloc[i,0]=E
            descriptor_molecule3d.iloc[i,1]=I
            descriptor_molecule3d.iloc[i,2]=R
            descriptor_molecule3d.iloc[i,3]=A
            descriptor_molecule3d.iloc[i,4]=S
            descriptor_molecule3d.iloc[i,5]=P1
            descriptor_molecule3d.iloc[i,6]=P2
            descriptor_molecule3d.iloc[i,7]=P3
            descriptor_molecule3d.iloc[i,8]=N1
            descriptor_molecule3d.iloc[i,9]=N2
        except RuntimeError:
            continue

descriptor_molecule3d=descriptor_molecule3d/100
descriptor_molecule3d

descriptor_molecule3d.isna().sum()

odor2=pd.concat([odor_descriptors, descriptor_molecule3d], axis=1)

odor2

odor2.to_json(r'C:\Users\nlupe\Desktop\QUART CURS\TFG\odor_descriptors.json', lines=False, orient='records')

Notas:
Con los Nans se podria hacer Multiple imputation
