==> 4 - MultipleImputation.ipynb <==
import pandas as pd
import numpy as np
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

odor = pd.read_json('odor_descriptors.json', lines=False, orient='records')

odor

odor.isna().sum()

odor_aux=odor.iloc[:,4:]
odor_aux

imp_mean = IterativeImputer(random_state=0)

imp_mean.fit(odor_aux)

odor_aux=imp_mean.transform(odor_aux)

odor_aux

len(odor_aux)

odor_imput = pd.DataFrame(odor_aux, columns=odor.columns[4:])

odor_imput

odor_final=pd.concat([odor.iloc[:,:4], odor_imput], axis=1)

odor_final.isna().sum()

odor_final.to_json(r'C:\Users\nlupe\Desktop\QUART CURS\TFG\odor_imputation.json', lines=False, orient='records')
