==> 5 - Word2Vec + KMeans.ipynb <==
import pandas as pd
from sklearn.cluster import KMeans
from sklearn.manifold import TSNE
import numpy as np
import matplotlib.pyplot as plt
#from wikipedia2vec import Wikipedia2Vec

odor = pd.read_json('Data/odor_relations.json', lines=False, orient='records')
odor_type=odor['name'].value_counts()
odor_type=odor_type[odor_type>38]
print(odor_type.index)
print(len(odor_type.index))

odors=odor_type.index
print(len(odors))

wiki2vec=Wikipedia2Vec.load('ModelFile')
memmaps=[]

def get_sentence_vector(word):
	m=0
	for i in word:
		m+=wiki2vec.get_word_vector(str(i))
		return m/len(word)

for i in odors:
	try:
		memmaps.append([i, wiki2vec.get_word_vector(str(i))])
	except:
		memmaps.append([i, get_sentence_vector(i)])

X=[]
for i in range(len(memmaps)):
    X.append(memmaps[i][1])
    
#pd.DataFrame(X).to_csv('X.csv')


tsne=TSNE(n_components=2)
df=tsne.fit_transform(X)

kmeans = KMeans(n_clusters=4, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 4 clusters')
plt.show()
plt.savefig('KMeans with 4 clusters.png')
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
pd.DataFrame(label0).to_csv('label4_0.csv')
pd.DataFrame(label1).to_csv('label4_1.csv')
pd.DataFrame(label2).to_csv('label4_2.csv')
pd.DataFrame(label3).to_csv('label4_3.csv')

kmeans = KMeans(n_clusters=5, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 5 clusters')
plt.show()
plt.savefig('KMeans with 5 clusters.png')
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
pd.DataFrame(label0).to_csv('label5_0.csv')
pd.DataFrame(label1).to_csv('label5_1.csv')
pd.DataFrame(label2).to_csv('label5_2.csv')
pd.DataFrame(label3).to_csv('label5_3.csv')
pd.DataFrame(label4).to_csv('label5_4.csv')

kmeans = KMeans(n_clusters=6, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 6 clusters')
plt.show()
plt.savefig('KMeans with 6 clusters.png')
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
pd.DataFrame(label0).to_csv('label6_0.csv')
pd.DataFrame(label1).to_csv('label6_1.csv')
pd.DataFrame(label2).to_csv('label6_2.csv')
pd.DataFrame(label3).to_csv('label6_3.csv')
pd.DataFrame(label4).to_csv('label6_4.csv')
pd.DataFrame(label5).to_csv('label6_5.csv')

kmeans = KMeans(n_clusters=7, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 7 clusters')
plt.show()
plt.savefig('KMeans with 7 clusters.png')
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
pd.DataFrame(label0).to_csv('label7_0.csv')
pd.DataFrame(label1).to_csv('label7_1.csv')
pd.DataFrame(label2).to_csv('label7_2.csv')
pd.DataFrame(label3).to_csv('label7_3.csv')
pd.DataFrame(label4).to_csv('label7_4.csv')
pd.DataFrame(label5).to_csv('label7_5.csv')
pd.DataFrame(label6).to_csv('label7_6.csv')


kmeans = KMeans(n_clusters=8, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 8 clusters')
plt.show()
plt.savefig('KMeans with 8 clusters.png')
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
label7 = odors[label == 7]
pd.DataFrame(label0).to_csv('label8_0.csv')
pd.DataFrame(label1).to_csv('label8_1.csv')
pd.DataFrame(label2).to_csv('label8_2.csv')
pd.DataFrame(label3).to_csv('label8_3.csv')
pd.DataFrame(label4).to_csv('label8_4.csv')
pd.DataFrame(label5).to_csv('label8_5.csv')
pd.DataFrame(label6).to_csv('label8_6.csv')
pd.DataFrame(label7).to_csv('label8_7.csv')

# Ahora, hace falta mirar de todos los grupos que se han creado, si el arquetipo químico concuerda entre ellos.

kmeans = KMeans(n_clusters=9, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 9 clusters')
plt.show()
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
label7 = odors[label == 7]
label8 = odors[label == 8]
pd.DataFrame(label0).to_csv('label9_0.csv')
pd.DataFrame(label1).to_csv('label9_1.csv')
pd.DataFrame(label2).to_csv('label9_2.csv')
pd.DataFrame(label3).to_csv('label9_3.csv')
pd.DataFrame(label4).to_csv('label9_4.csv')
pd.DataFrame(label5).to_csv('label9_5.csv')
pd.DataFrame(label6).to_csv('label9_6.csv')
pd.DataFrame(label7).to_csv('label9_7.csv')
pd.DataFrame(label8).to_csv('label9_8.csv')

kmeans = KMeans(n_clusters=10, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 10 clusters')
plt.show()
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
label7 = odors[label == 7]
label8 = odors[label == 8]
label9 = odors[label == 9]
pd.DataFrame(label0).to_csv('label10_0.csv')
pd.DataFrame(label1).to_csv('label10_1.csv')
pd.DataFrame(label2).to_csv('label10_2.csv')
pd.DataFrame(label3).to_csv('label10_3.csv')
pd.DataFrame(label4).to_csv('label10_4.csv')
pd.DataFrame(label5).to_csv('label10_5.csv')
pd.DataFrame(label6).to_csv('label10_6.csv')
pd.DataFrame(label7).to_csv('label10_7.csv')
pd.DataFrame(label8).to_csv('label10_8.csv')
pd.DataFrame(label9).to_csv('label10_9.csv')

kmeans = KMeans(n_clusters=11, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 11 clusters')
plt.show()
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
label7 = odors[label == 7]
label8 = odors[label == 8]
label9 = odors[label == 9]
label10 = odors[label == 10]
pd.DataFrame(label0).to_csv('label11_0.csv')
pd.DataFrame(label1).to_csv('label11_1.csv')
pd.DataFrame(label2).to_csv('label11_2.csv')
pd.DataFrame(label3).to_csv('label11_3.csv')
pd.DataFrame(label4).to_csv('label11_4.csv')
pd.DataFrame(label5).to_csv('label11_5.csv')
pd.DataFrame(label6).to_csv('label11_6.csv')
pd.DataFrame(label7).to_csv('label11_7.csv')
pd.DataFrame(label8).to_csv('label11_8.csv')
pd.DataFrame(label9).to_csv('label11_9.csv')
pd.DataFrame(label10).to_csv('label11_10.csv')

kmeans = KMeans(n_clusters=12, random_state=0).fit(X)
label = kmeans.labels_
index = np.unique(label)
for i in index:
    plt.scatter(df[label == i, 0], df[label == i, 1], label=i)
plt.legend()
plt.title('KMeans with 12 clusters')
plt.show()
label0 = odors[label == 0]
label1 = odors[label == 1]
label2 = odors[label == 2]
label3 = odors[label == 3]
label4 = odors[label == 4]
label5 = odors[label == 5]
label6 = odors[label == 6]
label7 = odors[label == 7]
label8 = odors[label == 8]
label9 = odors[label == 9]
label10 = odors[label == 10]
label11 = odors[label == 11]
pd.DataFrame(label0).to_csv('label12_0.csv')
pd.DataFrame(label1).to_csv('label12_1.csv')
pd.DataFrame(label2).to_csv('label12_2.csv')
pd.DataFrame(label3).to_csv('label12_3.csv')
pd.DataFrame(label4).to_csv('label12_4.csv')
pd.DataFrame(label5).to_csv('label12_5.csv')
pd.DataFrame(label6).to_csv('label12_6.csv')
pd.DataFrame(label7).to_csv('label12_7.csv')
pd.DataFrame(label8).to_csv('label12_8.csv')
pd.DataFrame(label9).to_csv('label12_9.csv')
pd.DataFrame(label10).to_csv('label12_10.csv')
pd.DataFrame(label11).to_csv('label12_11.csv')

odor
