==> 6 - AddKMeans8.ipynb <==
import pandas as pd

odor = pd.read_json('Data/odor_imputation.json', lines=False, orient='records')
odor_type=odor['name'].value_counts()
odor_type=odor_type[odor_type>38]
names_yes=odor_type.index

df = odor.loc[odor['name'].isin(names_yes)]
df.reset_index(drop=True, inplace=True)
df.head()

data=pd.read_csv('Data/classifyre_output.csv', on_bad_lines='skip', names=['Molecule Name', 'InChIKey', 'SMILES', 'Kingdom', 'Superclass', 'Class', 'Subclass', 'Intermediate Nodes', 'Direct Parent', 'Alternative Parents', 'Molecular Framework', 'Substituents', 'Structure-based description', 'Ancestors', 'External Descriptors'])

label0 = list(pd.read_csv('Data/LabelsMostFrequency/label8_0.csv')['0'])
label1 = list(pd.read_csv('Data/LabelsMostFrequency/label8_1.csv')['0'])
label2 = list(pd.read_csv('Data/LabelsMostFrequency/label8_2.csv')['0'])
label3 = list(pd.read_csv('Data/LabelsMostFrequency/label8_3.csv')['0'])
label4 = list(pd.read_csv('Data/LabelsMostFrequency/label8_4.csv')['0'])
label5 = list( pd.read_csv('Data/LabelsMostFrequency/label8_5.csv')['0'])
label6 = list( pd.read_csv('Data/LabelsMostFrequency/label8_6.csv')['0'])
label7 = list( pd.read_csv('Data/LabelsMostFrequency/label8_7.csv')['0'])

label = []
for i in range(len(df)):
    name=df['name'][i]
    if name in label0:
        label.append('0')
    elif name in label1:
        label.append(1)
    elif name in label2:
        label.append(2)
    elif name in label3:
        label.append(3)
    elif name in label4:
        label.append(4)
    elif name in label5:
        label.append(5)
    elif name in label6:
        label.append(6)
    elif name in label7:
        label.append(7)

df.insert(3, "label", label)

df.head()

prop=pd.DataFrame(columns=['Molecule Name', 'InChIKey', 'SMILES', 'Kingdom', 'Superclass', 'Class', 'Subclass', 'Intermediate Nodes', 'Direct Parent', 'Alternative Parents', 'Molecular Framework', 'Substituents', 'Structure-based description', 'Ancestors', 'External Descriptors'], index=range(0,13870))

for i in range(3787):
    for j in range(13870):
        if data['SMILES'][i].strip()==df['smiles'][j]:
            prop.loc[j] = data.loc[i]

prop.isna().sum()

odor_prop=pd.concat([df, prop], axis=1)

odor_prop.head()

odor_prop.to_json(r'C:\Users\nlupe\Desktop\QUART CURS\TFG\Data\odor_prop.json', lines=False, orient='records')
