==> 7 - AnalyseKMeans8.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numoy as np

odor = pd.read_json('Data/odor_prop.json', lines=False, orient='records')

#Subclass
for i in range(8):
    print('*********** Grupo', i, '***********')
    sc=odor[odor['label']==i]['Subclass'].value_counts()
    print(sc)
    scy=sc[0:10]
    scx=sc[0:10]
    sns.barplot(x=scy, y=scx.index)
    a = 'Grup ' + str(i)
    plt.title(a)
    plt.show()

#Class
for i in range(8):
    print('*********** Grupo', i, '***********')
    sc=odor[odor['label']==i]['Class'].value_counts()
    print(sc)
    scy=sc[0:10]
    scx=sc[0:10]
    sns.barplot(x=scy, y=scx.index)
    a = 'Grup ' + str(i)
    plt.title(a)
    plt.show()

#Superclass
for i in range(8):
    print('*********** Grupo', i, '***********')
    sc=odor[odor['label']==i]['Superclass'].value_counts()
    print(sc)
    scy=sc[0:10]
    scx=sc[0:10]
    sns.barplot(x=scy, y=scx.index)
    a = 'Grup ' + str(i)
    plt.title(a)
    plt.show()

#Direct parent
for i in range(8):
    print('*********** Grupo', i, '***********')
    sc=odor[odor['label']==i]['Direct Parent'].value_counts()
    print(sc)
    scy=sc[0:15]
    scx=sc[0:15]
    sns.barplot(x=scy, y=scx.index)
    a = 'Grup ' + str(i)
    plt.title(a)
    plt.show()

#Alternative Parents - pre
ap=odor['Alternative Parents'].unique()
ap_num=np.arange(0,234)
for i in range(len(odor)): 
    for j in range(len(ap)):
        if (odor['Alternative Parents'][i] == ap[j]):
            odor['Alternative Parents'][i] = ap_num[j]
#borramos 1 que son los none
odor=odor[odor['Alternative Parents'] != 1]

#borramos los 11 que son ' 1'
odor=odor[odor['Alternative Parents'] != 11]

#borramos los 102 que son ' '
odor=odor[odor['Alternative Parents'] != 102]

#Alternative Parents
groups = []
scxall = []

for i in range(8):
    print('*********** Grupo', i, '***********')
    sc=odor[odor['label']==i]['Alternative Parents'].value_counts()
    print(sc)
    scy=sc[0:15]
    scx=sc[0:15]
    print(scx)
    sns.barplot(x=scx.index, y=scy)
    a = 'Grup ' + str(i)
    plt.title(a)
    plt.show()
    for i in scx.index:
        scxall.append(i)

miss=[]
ind=np.unique(scxall)
for i in ind:
    miss.append(ap[i])
legend=pd.DataFrame()
legend['ind']=ind
legend['miss']=miss
legend.to_csv('Data/legend_alternative_parents.csv', index=False, sep=".")
