==> 8 - Classyfire Test 6.ipynb <==
Multiple imputation del test anterior

import pandas as pd
from sklearn import svm
from sklearn.model_selection import train_test_split
import seaborn as sns
import numpy as np
from scipy import stats
import matplotlib.pyplot as plt

from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVC

from sklearn.metrics import accuracy_score, log_loss
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

#### Tornem a carregar data pel primer nivell

odor = pd.read_json('Data/odor_prop.json', lines=False, orient='records')
odor=odor[odor['level']=='first']
odor.reset_index(drop=True, inplace=True)
mol=odor['cid'].value_counts()
mols=mol.index[mol>1]
groups=[]
for i in range(len(odor)):
    for mol in mols:
        if odor['cid'][i] == mol:
            g=np.max(odor[odor['cid']==mol]['label'])
            groups.append([mol, g])
            ls_drop=[]
for mol in groups:
    for i in range(len(odor)):
        if (odor['cid'][i] == mol[0]):
            if odor['label'][i]!=mol[1]:
                ls_drop.append(i)
odor=odor.drop(ls_drop)
odor.reset_index(drop=True, inplace=True)
odor = odor.sort_values(by='cid')
odor.reset_index(inplace=True)
ls_repeat=[]
for i in range(len(odor)-1):
    if (odor['cid'][i]==odor['cid'][i+1]):
        ls_repeat.append(i)
odor=odor.drop(ls_repeat)

#### Canviem string categoric a int

len(odor['Subclass'].unique()), len(odor['Superclass'].unique()), len(odor['Class'].unique()), len(odor['Direct Parent'].unique())

odor.replace(to_replace=odor['Subclass'].unique(), 
                   value=np.arange(0,69), inplace=True)
odor.replace(to_replace=odor['Superclass'].unique(), 
                   value=np.arange(0,10), inplace=True)
odor.replace(to_replace=odor['Class'].unique(), 
                   value=np.arange(0,65), inplace=True)
odor.replace(to_replace=odor['Direct Parent'].unique(), 
                   value=np.arange(0,113), inplace=True)

ind_X=['TPSA', 'MolLogP',
       'MolWt', 'ExactMolWt', 'MolMR', 'NumHAcceptors', 'NumHDonors',
       'NumRotatableBonds', 'NumAromaticRings', 'NumSaturatedRings',
       'NumAliphaticRings', 'NumAromaticHeterocycles',
       'NumSaturatedHeterocycles', 'NumAliphaticHeterocycles',
       'NumAromaticCarbocycles', 'NumSaturatedCarbocycles',
       'NumAliphaticCarbocycles', 'NumHeteroatoms', 'NumValenceElectrons',
       'RingCount', 'FractionCSP3', 'LabuteASA', 'BalabanJ', 'BertzCT', 'Ipc',
       'HallKierAlpha', 'Kappa1', 'Kappa2', 'Kappa3', 'Chi0', 'Chi1', 'Chi0n',
       'Chi1n', 'Chi2n', 'Chi3n', 'Chi4n', 'Chi0v', 'Chi1v', 'Chi2v', 'Chi3v',
       'Chi4v', 'HeavyAtom', 'Gasteiger', 'Eccentricity',
       'InertialShapeFactor', 'RadiusOfGyration', 'Asphericity',
       'SpherocityIndex', 'PMI1', 'PMI2', 'PMI3', 'NPR1', 'NPR2',
       'Superclass', 'Class', 'Subclass', 'Direct Parent'
      ]

#### Fem multiple imputation 

odor_aux = odor[ind_X]

from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

imp_mean = IterativeImputer(random_state=0)
imp_mean.fit(odor_aux)

odor_aux=imp_mean.transform(odor_aux)

odor_imput = pd.DataFrame(odor_aux, columns=ind_X)

odor_imput

odor_imput

ind_X=odor.columns[6:59]
X = odor_imput
y = odor['label']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=42) 

classifiers = [
    KNeighborsClassifier(3),
    DecisionTreeClassifier(),
    RandomForestClassifier(),
    LinearDiscriminantAnalysis(),
    make_pipeline(StandardScaler(), SVC(gamma='auto')),
    svm.SVC()
    ]

# Logging for Visual Comparison


for clf in classifiers:
    clf.fit(X_train, y_train)
    name = clf.__class__.__name__
    
    print("="*30)
    print(name)
    
    print('****Results****')
    train_predictions = clf.predict(X_test)
    acc = accuracy_score(y_test, train_predictions)
    print("Accuracy: {:.4%}".format(acc))
      
print("="*30)
